+++ tweaks/panel/src/SCIPanelRoot.m
@@ -8,7 +8,7 @@
 #import <objc/message.h>
 #import <objc/runtime.h>
 
-NSString *SCIVersionString = @"v0.6.1";  // AlbrhiPanel
+NSString *SCIVersionString = @"v0.6.2";  // AlbrhiPanel
 
 ///
 /// Albrhi's own control panel, in the iOS Settings app.